requirements.txt
# MedExamAI - Python Dependencies
# ================================

# Core dependencies (required)
pypdf>=3.17.0              # PDF text extraction
python-docx>=1.1.0         # DOCX document processing
pyyaml>=6.0.1              # Config file parsing
numpy>=1.24.0              # Numerical operations for embeddings
requests>=2.31.0           # HTTP requests for guideline fetching
anthropic                  # Anthropic API
boto3                      # AWS Bedrock client
python-dotenv              # Load .env secrets
biopython                  # PubMed/sequence utilities
bioservices                # Bio APIs
datacommons-pandas         # DataCommons access
tiktoken                   # Token counting
tenacity                   # Retry helper

# RAG System (for semantic search)
# openai>=1.3.0            # OpenAI API (optional, for embeddings)

# Optional - OCR for scanned PDFs
# pytesseract>=0.3.10      # Uncomment if needed

# Development dependencies (optional)
pytest>=7.4.3              # Unit testing
pytest-cov>=4.1.0          # Test coverage
# black>=23.12.0           # Code formatting
# pylint>=3.0.3            # Code linting
# ruff>=0.1.8              # Fast Python linter
